refactor (zigpyThread): split into four focused modules Extract send/concurrency logic → zigpySend.py, worker loop and command dispatch → workerLoop.py, radio config and app startup → radioStart.py, and HA-grade supervisor/watchdog/heartbeat → supervisor.py. zigpyThread.py is now a ~140-line thread entry-point (stop/start/setup/ thread-function) that imports _supervisor and _cleanup from supervisor.py. cleanup_unused_concurrency_state is re-exported from zigpySend for backward compatibility with Transport.py. AppGeneric.py updated to import zigpy_heartbeat_activity from supervisor.
