+++ Classes/ZigpyTransport/zigpyThread.py
@@ -11,80 +11,38 @@
 # SPDX-License-Identifier:    GPL-3.0 license
 
 """
-Zigbee Transport Module for Domoticz Plugin.
+zigpyThread.py — Thread entry-point for the Zigpy stack.
 
-This module provides the core functionality for handling Zigbee communication
-using the Zigpy library within the Domoticz plugin framework. It manages the
-Zigpy event loop, radio configuration, command dispatching, and transmission
-of Zigbee messages (unicast, multicast, broadcast). It supports multiple radio
-modules including EZSP (Bellows), ZNP, deCONZ, and BLZ.
+This is a thin orchestration layer.  It creates the OS thread, boots an
+asyncio event loop inside it, and hands control to the supervisor.
 
-Key features:
-- Asynchronous event loop for non-blocking operations.
-- Configuration setup for different radio backends.
-- Command processing and error handling with retries.
-- Concurrency limiting per device to prevent overload.
-- Integration with Domoticz via queues for sending/receiving data.
+All non-trivial logic lives in the modules imported below:
 
-Dependencies:
-- zigpy: Core Zigbee stack.
-- bellows: For EZSP (EmberZNet) support.
-- zigpy_znp: For ZNP (TI CC2531) support.
-- zigpy_deconz: For deCONZ support.
-- asyncio: For asynchronous operations.
-- queue: For thread-safe communication with Domoticz.
+    supervisor.py   — HA-grade restart supervisor, watchdog, heartbeat
+    radioStart.py   — radio config, app startup, post-startup frames
+    workerLoop.py   — writer-queue consumer, command dispatch
+    zigpySend.py    — Zigbee send helpers, concurrency limiting
 
-Configuration:
-- Relies on pluginconf for settings like channel, extended PAN ID, and radio module.
-- Uses logging via self.log for debugging and status updates.
-
-Usage:
-- Initialize via start_zigpy_thread(self).
-- Send commands via writer_queue with JSON-formatted data.
-- Receive responses via forwarder_queue.
+Public entry points consumed by Transport.py:
+    start_zigpy_thread(self)
+    stop_zigpy_thread(self)
+    cleanup_unused_concurrency_state(self)   # re-exported from zigpySend
 """
 
 import asyncio
-import contextlib
-import functools
-import importlib
-import json
-import queue
 import random
 import sys
 import time
-import traceback
-from pathlib import Path
 from threading import Thread
 
-import zigpy.config
-import zigpy.device
-import zigpy.types as t
-import zigpy.zcl
-from zigpy.exceptions import (APIException, ControllerException, DeliveryError,
-                              InvalidResponse)
-from zigpy_znp.exceptions import (CommandNotRecognized, InvalidCommandResponse,
-                                  InvalidFrame)
+from Classes.ZigpyTransport.supervisor import _cleanup, _supervisor
+from Classes.ZigpyTransport.zigpySend import \
+    cleanup_unused_concurrency_state  # noqa: F401  (re-export for Transport.py)
 
-from Classes.ZigpyTransport.plugin_encoders import (
-    build_plugin_0302_frame_content, build_plugin_8009_frame_content,
-    build_plugin_8011_frame_content,
-    build_plugin_8043_frame_list_node_descriptor,
-    build_plugin_8045_frame_list_controller_ep)
-from Classes.ZigpyTransport.tools import handle_thread_error
-from Modules.macPrefix import DELAY_FOR_VERY_KEY
 
-ERROR_TASK_CREATION_FAILED = 0xB6
-SEMAPHORE_TIMEOUT = 240  # seconds
-REQUEST_TIMEOUT = 8   # This is a given time for the request to be sent
-WAITING_TIME_BETWEEN_REQUESTS = .100
-MAX_CONCURRENT_REQUESTS_PER_DEVICE = 1
-VERIFY_KEY_DELAY = 6
-
-STARTUP_TIMEOUT = 120  # 2 min to receive the first heartbeat
-DEAD_STACK_THRESHOLD = 240   # seconds without heartbeat to consider the stack dead
-WATCH_DG_HEARTBEAT_INTERVAL = 10  # seconds between heartbeat checks (internal tick)
-MAX_RESTARTS_PER_HOUR = 5   # circuit-breaker: escalate to plugin restart after this many
+# ---------------------------------------------------------------------------
+# Thread lifecycle
+# ---------------------------------------------------------------------------
 
 def stop_zigpy_thread(self):
     """
@@ -125,7 +83,6 @@ def start_zigpy_thread(self):
     Sets the appropriate event loop policy for Windows compatibility and
     initializes the thread via setup_zigpy_thread if necessary.
     """
-
     self.log.logging("TransportZigpy", "Debug", "start_zigpy_thread - Starting Zigpy thread")
 
     # Set appropriate event loop policy for Windows compatibility
@@ -200,2160 +157,3 @@ def zigpy_thread_function(self):
         loop.run_forever()
     finally:
         _cleanup(self, loop)
-
-
-async def _supervisor(self):
-    """
-    High-level lifecycle supervisor for the Zigbee stack.
-
-    Implements a resilient restart strategy similar to Home Assistant / ZHA.
-
-    Responsibilities:
-        - Start and monitor the Zigbee stack lifecycle
-        - Detect crash or abnormal exit conditions
-        - Apply exponential backoff between restarts
-        - Track stack health state transitions
-        - Trigger optional radio recovery after repeated failures
-        - Escalate to a full plugin restart if the circuit breaker trips
-
-    State transitions:
-        STARTING → RUNNING → (EXIT | CRASHED) → RESTARTING
-
-    Restart policy:
-        - Exponential backoff: 2 s → 4 s → … → 60 s (capped)
-        - Radio recovery hook after every 5th consecutive restart
-        - Circuit breaker: escalate to plugin restart after MAX_RESTARTS_PER_HOUR
-          restarts within a rolling 1-hour window
-
-    Exit conditions:
-        - _zigpy_stop_requested is True  (external stop via stop_zigpy_thread)
-        - Circuit breaker tripped and plugin restart was requested
-
-    Design note — why _zigpy_stop_requested instead of _shutdown_event:
-        _run_zigbee_stack() *always* calls _shutdown_event.set() before it
-        returns, so that all intra-cycle tasks (watchdog, heartbeat, shutdown
-        sentinel) are cancelled cleanly.  If the supervisor loop tested
-        _shutdown_event it would exit after the very first run and never
-        restart.  _zigpy_stop_requested is a plain bool set only by
-        stop_zigpy_thread(), which allows the supervisor to distinguish a
-        genuine external shutdown from an internal restart trigger.
-        A fresh asyncio.Event is created at the start of every cycle so there
-        is no risk of a stale set() from a previous run bleeding through.
-    """
-    self.log.logging("TransportZigpy", "Log", "Supervisor started")
-
-    restart_delay = 2
-
-    while not self._zigpy_stop_requested:
-
-        self.log.logging(
-            "TransportZigpyStack", "Debug",
-            f"Supervisor: starting cycle #{self._restart_count} "
-            f"(consecutive_failures={self._consecutive_failures}, "
-            f"restart_delay={restart_delay}s, stop_requested={self._zigpy_stop_requested})"
-        )
-
-        # Fresh event for this run cycle — prevents a stale set() from a
-        # previous cycle from prematurely exiting the next run.
-        self._shutdown_event = asyncio.Event()
-
-        # Reset liveness tracking so the watchdog startup window is accurate
-        # for this cycle (stale timestamp from a prior run would suppress it).
-        self._last_heartbeat = None
-
-        # Clean up state left over from the previous run.
-        await _prepare_for_restart(self)
-
-        run_start = self.zigpy_loop.time()
-        self.log.logging("TransportZigpyStack", "Debug", f"Supervisor: entering _run_zigbee_stack (cycle #{self._restart_count})")
-
-        try:
-            self._stack_health = "STARTING"
-            await _run_zigbee_stack(self)
-            self.log.logging("TransportZigpy", "Log", "Zigbee stack exited")
-
-        except Exception as e:
-            self.log.logging("TransportZigpy", "Error", f"Zigbee crash: {e}")
-            self._stack_health = "CRASHED"
-
-        run_duration = self.zigpy_loop.time() - run_start
-        self.log.logging(
-            "TransportZigpyStack", "Debug",
-            f"Supervisor: _run_zigbee_stack returned after {run_duration:.1f}s, "
-            f"health={self._stack_health}, stop_requested={self._zigpy_stop_requested}"
-        )
-
-        # If the run ended because stop_zigpy_thread() was called, leave now
-        # without scheduling a restart.
-        if self._zigpy_stop_requested:
-            self.log.logging("TransportZigpy", "Log", "Supervisor: stop requested — not restarting")
-            break
-
-        # --- restart / recovery logic -----------------------------------------
-
-        self._restart_count += 1
-        self._consecutive_failures += 1
-
-        # Stability reset: if the stack ran for > 5 min, the crash is likely
-        # a transient event — reset backoff and consecutive failure counter so
-        # the next restart is fast and radio recovery is not triggered early.
-        if run_duration > 300:
-            self.log.logging(
-                "TransportZigpy", "Log",
-                f"Supervisor: stack was stable for {run_duration:.0f}s — resetting backoff and failure counter"
-            )
-            restart_delay = 2
-            self._consecutive_failures = 0
-
-        # Circuit breaker: count restarts in the last hour.
-        now = time.monotonic()
-        self._restart_timestamps = [t for t in self._restart_timestamps if now - t < 3600]
-        self._restart_timestamps.append(now)
-
-        self.log.logging(
-            "TransportZigpyStack", "Debug",
-            f"Supervisor: circuit-breaker count={len(self._restart_timestamps)}/{MAX_RESTARTS_PER_HOUR} in last hour"
-        )
-        if len(self._restart_timestamps) >= MAX_RESTARTS_PER_HOUR:
-            self.log.logging(
-                "TransportZigpy", "Error",
-                f"Supervisor: {MAX_RESTARTS_PER_HOUR} restarts in 1 h — escalating to plugin restart"
-            )
-            if callable(getattr(self, "restart_plugin", None)):
-                self.restart_plugin()
-            break
-
-        # Radio recovery after repeated *consecutive* failures (not total count).
-        if self._consecutive_failures > 0 and self._consecutive_failures % 5 == 0:
-            self.log.logging(
-                "TransportZigpy", "Error",
-                f"Supervisor: {self._consecutive_failures} consecutive failures — attempting radio recovery"
-            )
-            await _maybe_reset_radio(self)
-
-        self.log.logging(
-            "TransportZigpy", "Log",
-            f"Supervisor: restarting in {restart_delay}s "
-            f"(attempt #{self._restart_count}, consecutive={self._consecutive_failures})"
-        )
-        self.log.logging(
-            "TransportZigpyStack", "Debug",
-            f"Supervisor: sleeping {restart_delay}s before next cycle "
-            f"(next restart_delay will be {min(restart_delay * 2, 60)}s)"
-        )
-        await asyncio.sleep(restart_delay)
-        restart_delay = min(restart_delay * 2, 60)
-
-    # Signal the event loop to stop so zigpy_thread_function's run_forever()
-    # returns and _cleanup() is called.
-    self.log.logging("TransportZigpy", "Log", "Supervisor exiting — stopping event loop")
-    asyncio.get_running_loop().stop()
-
-
-async def _prepare_for_restart(self):
-    """
-    Reset volatile state before a supervised restart.
-
-    Called by _supervisor at the start of every cycle.  Ensures that a
-    crash or timeout in the previous run does not leave dangling resources
-    (open serial port, stale commands, locked semaphores) visible to the
-    next start_zigpy_task invocation.
-
-    Safe to call on the very first cycle (all attributes are None / empty).
-    """
-    # If the previous run's app was not shut down cleanly (e.g. radio_start
-    # timed out), force-disconnect now before re-opening the serial port.
-    if self.app is not None:
-        self.log.logging("TransportZigpyStack", "Debug", "_prepare_for_restart: app still set — force-disconnecting")
-        with contextlib.suppress(Exception):
-            await asyncio.wait_for(self.app.disconnect(), timeout=5.0)
-        self.app = None
-        self.log.logging("TransportZigpyStack", "Debug", "_prepare_for_restart: force-disconnect done")
-    else:
-        self.log.logging("TransportZigpyStack", "Debug", "_prepare_for_restart: app is None, no disconnect needed")
-
-    # Replace the writer queue so stale pre-crash commands are not replayed.
-    self.writer_queue = queue.Queue()
-    self.log.logging("TransportZigpyStack", "Debug", "_prepare_for_restart: fresh writer_queue created")
-
-    # Clear per-device concurrency tracking — semaphores from the previous
-    # run are invalid after a restart.
-    self._concurrent_requests_semaphores_list.clear()
-    self._currently_waiting_requests_list.clear()
-    self._currently_not_reachable.clear()
-
-    # Ensure the running flag is in a known state for the next start_zigpy_task.
-    self.zigpy_running = False
-    self.log.logging("TransportZigpyStack", "Debug", "_prepare_for_restart: state reset complete")
-
-
-async def _run_zigbee_stack(self):
-    """
-    Run the Zigbee stack as a set of concurrent tasks for one lifecycle.
-
-    Starts three tasks in parallel:
-        - zigbee_task:    the main radio loop (start_zigpy_task)
-        - watchdog_task:  health monitor and hard-failure detector
-        - shutdown_task:  waits for the external _shutdown_event signal
-
-    Returns as soon as *any* of them finishes (FIRST_COMPLETED), then
-    sets _shutdown_event and cancels the remaining tasks.
-
-    The supervisor treats any return from this coroutine as a signal to
-    restart, unless _zigpy_stop_requested is set.
-    """
-    self.log.logging("TransportZigpy", "Log", "_run_zigbee_stack starting")
-
-    zigbee_task = asyncio.create_task(start_zigpy_task(self, channel=0, extended_pan_id=0))
-    shutdown_task = asyncio.create_task(self._shutdown_event.wait())
-    watchdog_task = asyncio.create_task(_watchdog(self))
-
-    done, pending = await asyncio.wait(
-        {
-            zigbee_task, 
-            shutdown_task, 
-            watchdog_task
-            },
-        return_when=asyncio.FIRST_COMPLETED
-    )
-    self.log.logging("TransportZigpyStack", "Log", "_run_zigbee_stack - one task completed, initiating shutdown sequence")
-
-    # Log which task(s) triggered the completion for diagnostics
-    task_names = {zigbee_task: "zigbee_task", shutdown_task: "shutdown_task", watchdog_task: "watchdog_task"}
-    for _t in done:
-        exc = None
-        with contextlib.suppress(Exception):
-            exc = _t.exception()
-
-        self.log.logging(
-            "TransportZigpyStack", "Debug",
-            f"_run_zigbee_stack: '{task_names.get(_t, _t.get_name())}' completed "
-            f"(cancelled={_t.cancelled()}, exception={exc!r})"
-        )
-
-    self._shutdown_event.set()
-
-    for task in pending:
-        task.cancel()
-
-    await asyncio.gather(*pending, return_exceptions=True)
-
-    # All three lifecycle tasks (zigbee_task, shutdown_task, watchdog_task) are
-    # now done.  asyncio.current_task() here is the _supervisor coroutine itself,
-    # so it is automatically excluded.  Cancel any stray application tasks that
-    # survived (e.g. send_unicast_command-*, topology-scan, etc.).
-    lifecycle_tasks = {zigbee_task, shutdown_task, watchdog_task}
-    stray = [
-        t for t in asyncio.all_tasks()
-        if t is not asyncio.current_task() and t not in lifecycle_tasks
-    ]
-    if stray:
-        self.log.logging(
-            "TransportZigpyStack", "Debug",
-            f"_run_zigbee_stack: cancelling {len(stray)} stray task(s): "
-            f"{[t.get_name() for t in stray]}"
-        )
-        for t in stray:
-            t.cancel()
-        await asyncio.gather(*stray, return_exceptions=True)
-
-    self.log.logging("TransportZigpyStack", "Log", "_run_zigbee_stack ending")
-
-
-def _cleanup(self, loop):
-    """
-    Performs a safe shutdown of the asyncio event loop and all pending tasks.
-
-    This function ensures:
-        - All running tasks are cancelled
-        - Tasks are given opportunity to terminate gracefully
-        - Remaining coroutines are awaited to avoid warnings
-        - Event loop is safely closed
-
-    Purpose:
-        Prevents:
-            - Task leakage
-            - "coroutine was never awaited" warnings
-            - Zombie tasks after plugin shutdown
-
-    Warning:
-        This function assumes no new tasks are being created at shutdown time.
-    """
-
-    self.log.logging("TransportZigpy", "Log", "Cleaning up loop")
-
-    for task in asyncio.all_tasks(loop):
-        task.cancel()
-
-    try:
-        loop.run_until_complete(
-            asyncio.gather(*asyncio.all_tasks(loop), return_exceptions=True)
-        )
-    except Exception:
-        pass  # nosec B110 (nothing to do if happen)
-
-    loop.close()
-
-
-def zigpy_heartbeat_activity(self):
-    """
-    Records Zigbee activity to indicate that the stack is alive.
-
-    This function is called from Zigbee event handlers whenever
-    meaningful network activity occurs (e.g., incoming messages,
-    device updates, or command responses).
-
-    Responsibilities:
-        - Update last activity timestamp
-        - Update last heartbeat timestamp
-        - Indicate that Zigbee stack is actively processing traffic
-
-    Notes:
-        - This is NOT a time-based heartbeat
-        - It is an event-driven liveness indicator
-    """
-    if self.zigpy_loop is None:
-        return
-
-    now = self.zigpy_loop.time()
-
-    self._last_activity = now
-    self._last_heartbeat = now
-
-
-async def _watchdog(self):
-    """
-    Combined health monitor and hard-failure detector for the Zigbee stack.
-
-    Replaces the previous split between _heartbeat_monitor (observational)
-    and _watchdog (failure detection) with a single task that does both.
-
-    Health classification (updates _stack_health):
-        ALIVE   — last heartbeat < 30 s ago
-        IDLE    — last heartbeat 30–120 s ago
-        SUSPECT — last heartbeat 120 s – DEAD_STACK_THRESHOLD ago
-        DEAD    — last heartbeat > DEAD_STACK_THRESHOLD seconds ago
-
-    Failure conditions that trigger a supervised restart:
-        1. No heartbeat received within 120 s of startup (startup timeout).
-        2. Heartbeat gap exceeds DEAD_STACK_THRESHOLD (hard failure).
-
-    Logging policy:
-        - State transitions are logged at Log level.
-        - Failure triggers are logged at Error level.
-        - No per-tick log spam (previous version logged every 5 s).
-
-    Poll interval: WATCH_DG_HEARTBEAT_INTERVAL.
-    """
-    self.log.logging("TransportZigpy", "Log", "Watchdog started")
-
-    loop = self.zigpy_loop
-    startup_deadline = loop.time() + STARTUP_TIMEOUT  # 2 min to receive the first heartbeat
-    _prev_health = self._stack_health
-
-    while not self._shutdown_event.is_set():
-        await asyncio.sleep(WATCH_DG_HEARTBEAT_INTERVAL)
-        self.log.logging( "TransportZigpy", "Log", f"Watchdog tick: now: {loop.time()}, last_heartbeat={self._last_heartbeat}, startup_deadline={startup_deadline}, current_health={self._stack_health}" )
-
-        now = loop.time()
-
-        if self._last_heartbeat is None:
-            # Still waiting for the first heartbeat after (re)start.
-            if now > startup_deadline:
-                self.log.logging(
-                    "TransportZigpyStack", "Error",
-                    "Watchdog: no heartbeat within 120s startup window — triggering supervised restart"
-                )
-                self._stack_health = "DEAD"
-                self._shutdown_event.set()
-                return
-
-            self.log.logging(
-                "TransportZigpyStack", "Debug",
-                f"Watchdog: waiting for first heartbeat, {max(0, startup_deadline - now):.0f}s remaining in startup window"
-            )
-            continue
-
-        delta = now - self._last_heartbeat
-
-        # Classify current health.
-        if delta < 30:
-            new_health = "ALIVE"
-        elif delta < 120:
-            new_health = "IDLE"
-        elif delta < DEAD_STACK_THRESHOLD:
-            new_health = "SUSPECT"
-        else:
-            new_health = "DEAD"
-
-        # Log only on state transitions — no per-tick noise.
-        if new_health != _prev_health:
-            self.log.logging(
-                "TransportZigpyStack", "Log",
-                f"Watchdog: stack health {_prev_health} → {new_health} ({delta:.0f}s since last heartbeat)"
-            )
-            _prev_health = new_health
-            self._stack_health = new_health
-
-        if new_health == "DEAD":
-            self.log.logging(
-                "TransportZigpyStack", "Error",
-                f"Watchdog: heartbeat lost for {delta:.0f}s — triggering supervised restart"
-            )
-            self._shutdown_event.set()
-            return
-
-
-async def _maybe_reset_radio(self):
-    """
-    Last-resort radio recovery: force-disconnect the adapter and wait for
-    the OS to release the serial/USB port before the next restart cycle.
-
-    Invoked by the supervisor after every 5th consecutive failure.  The
-    actual reconnect is left to the next supervisor cycle's radio_start call.
-
-    Strategy:
-        1. Disconnect the zigpy ControllerApplication (closes serial/USB fd).
-        2. Null the app reference so _prepare_for_restart doesn't retry.
-        3. Sleep 5 s to give the OS time to fully release the port.
-
-    After this function returns the supervisor will call _prepare_for_restart
-    and then _run_zigbee_stack → radio_start as normal.
-    """
-    self.log.logging(
-        "TransportZigpy", "Error",
-        f"Radio recovery: forcing adapter disconnect "
-        f"(consecutive failures: {self._consecutive_failures})"
-    )
-    try:
-        if self.app is not None:
-            with contextlib.suppress(Exception):
-                await asyncio.wait_for(self.app.disconnect(), timeout=10.0)
-            self.app = None
-
-        self.log.logging(
-            "TransportZigpy", "Status",
-            "Radio recovery: adapter disconnected — waiting 5s for OS to release port"
-        )
-        await asyncio.sleep(5)
-        self.log.logging(
-            "TransportZigpy", "Status",
-            "Radio recovery: ready — next supervisor cycle will reconnect"
-        )
-    except Exception as e:
-        self.log.logging("TransportZigpy", "Error", f"Radio reset error: {e}")
-
-
-async def start_zigpy_task(self, channel, extended_pan_id):
-    """
-    Asynchronous task to start the Zigpy application and run the worker loop.
-
-    Configures channel and extended PAN ID from plugin config, starts the radio,
-    initializes the writer_queue, runs the worker_loop, and handles shutdown.
-    """
-    self.log.logging("TransportZigpy", "Debug", "start_zigpy_task - Starting zigpy stack with channel %s and extended PAN ID 0x%016x" % (channel, extended_pan_id))
-    self.zigpy_running = True
-    
-    if "channel" in self.pluginconf.pluginConf:
-        channel = int(self.pluginconf.pluginConf["channel"])
-
-    if "extendedPANID" in self.pluginconf.pluginConf:
-        if isinstance( self.pluginconf.pluginConf["extendedPANID"], str):
-            extended_pan_id = int(self.pluginconf.pluginConf["extendedPANID"], 16)
-        else:
-            extended_pan_id = self.pluginconf.pluginConf["extendedPANID"]
-
-    self.log.logging( "TransportZigpy", "Debug", f"start_zigpy_task -extendedPANID {self.pluginconf.pluginConf['extendedPANID']} {extended_pan_id}", )
-
-    try:
-        await asyncio.wait_for(
-            radio_start(self, self.statistics, self.pluginconf, self.use_of_zigpy_persistent_db, self._radiomodule, self._serialPort, set_channel=channel, set_extendedPanId=extended_pan_id),
-            timeout=60.0   # 60s to open serial and get network info
-        )
-    except asyncio.TimeoutError:
-        self.log.logging("TransportZigpy", "Error",
-                         "radio_start timed out after 60s — triggering supervised restart")
-        # asyncio.wait_for cancels radio_start via CancelledError, which may
-        # leave the serial port open.  Disconnect now so the OS releases the
-        # fd before the next restart cycle tries to reopen it.
-        if self.app is not None:
-            with contextlib.suppress(Exception):
-                await asyncio.wait_for(self.app.disconnect(), timeout=5.0)
-            self.app = None
-        return   # start_zigpy_task returns → supervisor restarts
-
-    except Exception as e:
-        self.log.logging("TransportZigpy", "Error", f"start_zigpy_task error in radio_start: {e}")
-
-    self.log.logging(
-        "TransportZigpyStack", "Debug",
-        f"start_zigpy_task: radio_start finished, app={'set' if self.app else 'None'}, zigpy_running={self.zigpy_running}"
-    )
-
-    # Run forever
-    self.writer_queue = queue.Queue()  # We MUST use queue and not asyncio.Queue, because it is not compatible with the Domoticz framework
-
-    try:
-        await worker_loop(self)
-
-    except asyncio.CancelledError:
-        # Handle cancellation gracefully if the loop is stopped externally
-        self.log.logging("TransportZigpy", "Error", "start_zigpy_task worker_loop(self) was cancelled.")
-
-    except RuntimeError as e:
-        # Handle cases like "Cannot run the event loop while another loop is running"
-        self.log.logging("TransportZigpy", "Error", f"start_zigpy_task worker_loop(self) encountered a runtime error: {e}")
-        
-    except Exception as e:
-        # Log any errors encountered during execution
-        self.log.logging("TransportZigpy", "Error", f"start_zigpy_task worker_loop(self) error: {e}")
-
-    # We exit the worker_loop, shutdown time
-    # worker_loop has exited — clear the supervisor flag before shutdown
-    self.log.logging("TransportZigpyStack", "Debug", f"start_zigpy_task: worker_loop exited, zigpy_running={self.zigpy_running}")
-    if self.app:
-        self.app._supervisor_running = False
-        self.log.logging("TransportZigpyStack", "Debug", "start_zigpy_task: _supervisor_running cleared on app")
-
-    try:
-        self.log.logging("TransportZigpy", "Debug", "Shutting down zigpy thread...")
-        if self.app:
-            await self.app.shutdown()
-
-    except Exception as e:
-        self.log.logging("TransportZigpy", "Error", f"start_zigpy_task shutdown(self) error: {e}")
-        self.log.logging("TransportZigpy", "Error", f" {str(traceback.format_exc())}")
-
-        self.log.logging("TransportZigpy", "Log", "Disconnecting communication")
-        if self.app:
-            await self.app.disconnect()
-
-    await asyncio.sleep(1)
-
-    self.log.logging(["TransportZigpy", "StopProcess"], "Debug", "start_zigpy_task - exiting zigpy task.")
-
-
-
-async def radio_start(self, statistics, pluginconf, use_of_zigpy_persistent_db, radiomodule, serialPort, auto_form=False, set_channel=0, set_extendedPanId=0):
-    """
-    Initializes the Zigpy application for the specified radio module.
-
-    Sets up configuration based on the radio module (EZSP, ZNP, deCONZ, BLZ),
-    applies optional configurations, starts the application, loads persistent
-    DB if enabled, and performs post-startup actions.
-
-    Args:
-        statistics: Statistics tracker object.
-        pluginconf: Plugin configuration object.
-        use_of_zigpy_persistent_db: Flag to enable persistent DB.
-        radiomodule (str): Type of radio module ('ezsp', 'znp', 'deCONZ', 'blz').
-        serialPort (str): Serial port path for the radio.
-        auto_form (bool): Whether to auto-form the network.
-        set_channel (int): Channel to set for the network.
-        set_extendedPanId (int): Extended PAN ID to set.
-
-    Returns:
-        None
-    """
-    self.log.logging("TransportZigpy", "Debug", "In radio_start %s" %radiomodule)
-    config = None
-
-    serial_specifics = self._serialPort_communication_specifics or {}
-
-    _RADIO_REGISTRY = {
-        "ezsp": ("bellows.config", "Classes.ZigpyTransport.AppBellows.App_bellows", ezsp_configuration_setup),
-        "znp": ("zigpy_znp.config", "Classes.ZigpyTransport.AppZnp.App_znp", znp_configuration_setup),
-        "deCONZ":("zigpy_deconz.config", "Classes.ZigpyTransport.AppDeconz.App_deconz", deconz_configuration_setup),
-        "blz": (None, "Classes.ZigpyTransport.AppBlz.App_blz", blz_configuration_setup),
-    }
-
-    entry = _RADIO_REGISTRY.get(radiomodule)
-    if entry is None:
-        self.log.logging("TransportZigpy", "Error", f"Wrong radiomodule: {radiomodule}")
-        return
-
-    conf_module_path, app_class_path, setup_fn = entry
-    radio_specific_conf = importlib.import_module(conf_module_path) if conf_module_path else {}
-    App = _import_class(app_class_path)
-    config = setup_fn(self, radio_specific_conf, serialPort, serial_specifics)
-
-    self.log.logging("TransportZigpy", "Status", "++ Started radio %s port: %s config %s" %( radiomodule, serialPort, config))
-
-    try:
-        optional_configuration_setup(self, config, radio_specific_conf, set_extendedPanId, set_channel)
-
-    except Exception as e:
-        self.log.logging( "TransportZigpy", "Error", "Error while applying optional configuration to Radio: %s on port %s with %s" %( radiomodule, serialPort, e) )
-        self.log.logging("TransportZigpy", "Error", "%s" %traceback.format_exc())
-        return
-
-    try:
-        if radiomodule in ["znp", "deCONZ", "ezsp", "blz"]:
-            self.app = App(config)
-        else:
-            self.log.logging( "TransportZigpy", "Error", "Wrong radiomode: %s" % (radiomodule), )
-            return
-
-    except Exception as e:
-        self.log.logging( "TransportZigpy", "Error", "Error while starting radio %s on port: %s - Error: %s" %( radiomodule, serialPort, e) )
-        return
-
-    if self.pluginParameters["Mode3"] == "True":
-        self.log.logging( "TransportZigpy", "Status", "++ Coordinator initialisation requested Channel %s(0x%02x) ExtendedPanId: 0x%016x" % (
-            set_channel, set_channel, set_extendedPanId), )
-        new_network = True
-
-    else:
-        new_network = False
-
-    if self.app and self.use_of_zigpy_persistent_db:
-        self.log.logging( "TransportZigpy", "Status", "++ Use of Zigpy Persistent Db")
-        try:
-            await self.app._load_db()
-        except Exception as e:
-            self.log.logging( "TransportZigpy", "Error", "++ Error loading Zigpy Persistent Db: %s" %e)
-
-    try:
-        await _radio_startup(self, statistics, pluginconf, use_of_zigpy_persistent_db, new_network, radiomodule)
-
-    except Exception as e:
-        self.log.logging( "TransportZigpy", "Error", "Error during radio startup: %s" %e)
-
-    self.log.logging( "TransportZigpy", "Debug", "Exiting co-rounting radio_start")
-
-
-def ezsp_configuration_setup(self, bellows_conf, serialPort, serial_specifics):
-    """
-    Sets up the configuration dictionary for EZSP (Bellows) radio module.
-
-    Includes device path, baudrate, flow control, and optional policies like
-    unsecured rejoins, end device children limit, and TX power mode.
-
-    Args:
-        self: Instance of the transport class.
-        bellows_conf: Bellows configuration module.
-        serialPort (str): Serial port for the device.
-        serial_specifics (dict): Serial communication specifics.
-
-    Returns:
-        dict: Configuration dictionary for the EZSP application.
-    """
-    # The bellows implementation of flow control is a bit special
-    # if config[zigpy.config.CONF_DEVICE_FLOW_CONTROL] is None:
-    #     xon_xoff, rtscts = True, False
-    # else:
-    #     xon_xoff, rtscts = False, True
-
-    flow_control = serial_specifics.get("FlowControl", None)
-    if flow_control == "software":
-        flow_control = None
-
-    config = {
-        zigpy.config.CONF_DEVICE: {
-            zigpy.config.CONF_DEVICE_PATH: serialPort,
-            zigpy.config.CONF_DEVICE_BAUDRATE: serial_specifics.get("Baudrate", 115200),
-            zigpy.config.CONF_DEVICE_FLOW_CONTROL: flow_control
-        },
-        zigpy.config.CONF_NWK: {
-        },
-        bellows_conf.CONF_EZSP_CONFIG: {
-        },
-        # configure automatic behaviors in the NCP (Network Co-Processor)
-        bellows_conf.CONF_EZSP_POLICIES: {
-        },
-        zigpy.config.CONF_OTA: {
-        },
-        "handle_unknown_devices": True,
-    }
-
-    if self.pluginconf.pluginConf.get("EzspAllowUnsecuredRejoins"):
-        self.log.logging( "TransportZigpy", "Status", "++ Allow Unsecure Rejoins for Aqara devices ...")
-        # “If a device tries to rejoin without a secure link key, still let it in.”
-        config[bellows_conf.CONF_EZSP_POLICIES]["TRUST_CENTER_POLICY"] = 0x0003   # ALLOW_UNSECURED_REJOINS|ALLOW_JOINS
-          
-    if self.pluginconf.pluginConf.get("BellowsNoMoreEndDeviceChildren"):
-        self.log.logging("TransportZigpy", "Status", "++ Set The maximum number of end device children that Coordinater will support to 0")
-        config[bellows_conf.CONF_EZSP_CONFIG]["CONFIG_MAX_END_DEVICE_CHILDREN"] = 0
-
-    if self.pluginconf.pluginConf.get("TXpower_set") is not None:
-        tx_power = int(self.pluginconf.pluginConf["TXpower_set"])
-        self.log.logging("TransportZigpy", "Status", f"++ Setting TX Power to {tx_power} dBm")
-
-        # Boost mode flag (keep if you want PA boost enabled)
-        config[bellows_conf.CONF_EZSP_CONFIG]["CONFIG_TX_POWER_MODE"] = 0x3
-        self.log.logging("TransportZigpy", "Status", "++ Setting mode of transmission power adjustment 'TX_POWER_MODE_BOOST' dBm")
-
-        # Actual power level via zigpy network config
-        config[zigpy.config.CONF_NWK][zigpy.config.CONF_NWK_TX_POWER] = tx_power
-
-    return config
-
-
-def znp_configuration_setup(self, znp_conf, serialPort, serial_specifics):
-    """
-    Sets up the configuration dictionary for ZNP radio module.
-
-    Includes device path, baudrate, flow control, and optional settings like
-    endpoint preference and TX power.
-
-    Args:
-        self: Instance of the transport class.
-        znp_conf: ZNP configuration module.
-        serialPort (str): Serial port for the device.
-        serial_specifics (dict): Serial communication specifics.
-
-    Returns:
-        dict: Configuration dictionary for the ZNP application.
-    """
-    config = {
-        zigpy.config.CONF_DEVICE: {
-            zigpy.config.CONF_DEVICE_PATH: serialPort,
-            zigpy.config.CONF_DEVICE_BAUDRATE: serial_specifics.get("Baudrate", 115200),
-            zigpy.config.CONF_DEVICE_FLOW_CONTROL: serial_specifics.get("FlowControl", None)
-        },
-        zigpy.config.CONF_NWK: {
-        },
-        znp_conf.CONF_ZNP_CONFIG: {
-        },
-        zigpy.config.CONF_OTA: {
-        },
-    }
-    if self.pluginconf.pluginConf.get("TXpower_set") is not None:
-        tx_power = int(self.pluginconf.pluginConf["TXpower_set"])
-        config[znp_conf.CONF_ZNP_CONFIG]["tx_power"] = tx_power          # znp-specific
-        config[zigpy.config.CONF_NWK][zigpy.config.CONF_NWK_TX_POWER] = tx_power  # zigpy layer
-
-    if specific_endpoints(self):
-        config[ znp_conf.CONF_ZNP_CONFIG][ "prefer_endpoint_1" ] = False
-    
-        
-    return config
-
-
-def deconz_configuration_setup(self, deconz_conf, serialPort, serial_specifics):
-    """
-    Sets up the configuration dictionary for deCONZ radio module.
-
-    Basic configuration including device path, baudrate, and flow control.
-
-    Args:
-        self: Instance of the transport class.
-        deconz_conf: deCONZ configuration module.
-        serialPort (str): Serial port for the device.
-        serial_specifics (dict): Serial communication specifics.
-
-    Returns:
-        dict: Configuration dictionary for the deCONZ application.
-    """
-    return {
-        zigpy.config.CONF_DEVICE: {
-            zigpy.config.CONF_DEVICE_PATH: serialPort, 
-            zigpy.config.CONF_DEVICE_BAUDRATE: serial_specifics.get("Baudrate", 115200),
-            zigpy.config.CONF_DEVICE_FLOW_CONTROL: serial_specifics.get("FlowControl", None)
-        },
-        zigpy.config.CONF_NWK: {
-        },
-        zigpy.config.CONF_OTA: {
-        },
-    }
-
-
-def blz_configuration_setup(self, blz_conf, serialPort, serial_specifics):
-    """
-    Sets up the configuration dictionary for BLZ radio module.
-
-    Basic configuration including device path, baudrate (default 2M), and flow control.
-
-    Args:
-        self: Instance of the transport class.
-        blz_conf: BLZ configuration module.
-        serialPort (str): Serial port for the device.
-        serial_specifics (dict): Serial communication specifics.
-
-    Returns:
-        dict: Configuration dictionary for the BLZ application.
-    """
-    return {
-        zigpy.config.CONF_DEVICE: {
-            zigpy.config.CONF_DEVICE_PATH: serialPort, 
-            zigpy.config.CONF_DEVICE_BAUDRATE: serial_specifics.get("Baudrate", 2000000),
-            zigpy.config.CONF_DEVICE_FLOW_CONTROL: serial_specifics.get("FlowControl", None)
-        },
-        zigpy.config.CONF_NWK: {
-        },
-        zigpy.config.CONF_OTA: {
-        },
-    }
-
-
-def optional_configuration_setup(self, config, radio_conf, set_extendedPanId, set_channel):
-    """
-    Applies optional Zigpy configuration settings.
-
-    Sets extended PAN ID, channel, source routing, OTA, topology scan, watchdog,
-    database persistence, network backup, and startup energy scan based on plugin config.
-
-    Args:
-        self: Instance of the transport class.
-        config (dict): Base configuration dictionary to update.
-        radio_conf: Radio-specific configuration module.
-        set_extendedPanId (int): Extended PAN ID to set if non-zero.
-        set_channel (int): Channel to set if non-zero.
-    """
-    # In case we have to set the Extended PAN Id
-    if set_extendedPanId != 0:
-        config[zigpy.config.CONF_NWK][zigpy.config.CONF_NWK_EXTENDED_PAN_ID] = "%s" % ( t.EUI64(t.uint64_t(set_extendedPanId).serialize()) )
-
-    # In case we have to force the Channel
-    if radio_conf and set_channel != 0:
-        config[zigpy.config.CONF_NWK][zigpy.config.CONF_NWK_CHANNEL] = set_channel
-
-    if self.pluginconf.pluginConf.get("TXpower_set") is not None:
-        config[zigpy.config.CONF_NWK][zigpy.config.CONF_NWK_TX_POWER] = int(self.pluginconf.pluginConf["TXpower_set"])
-
-    # Enable or not Source Routing based on zigpySourceRouting setting
-    config[zigpy.config.CONF_SOURCE_ROUTING] = bool( self.pluginconf.pluginConf["zigpySourceRouting"] )
-    
-    # Disable Zigpy OTA
-    config[zigpy.config.CONF_OTA][zigpy.config.CONF_OTA_ENABLED] = False
-    
-    # Disable zigpy conf topo scan by default
-    config[zigpy.config.CONF_TOPO_SCAN_ENABLED] = False
-
-    # Enable Zigpy Watchdog by default
-    config[zigpy.config.CONF_WATCHDOG_ENABLED] = True
-
-    # Reduce the number of simultaneous APS transactions    
-    config[zigpy.config.CONF_MAX_CONCURRENT_REQUESTS] = 4
-
-    # Config Zigpy db. if not defined, there is no persistent Db.
-    if self.pluginconf.pluginConf.get("enableZigpyPersistentInFile"):
-        data_folder = Path( self.pluginconf.pluginConf["pluginData"] )
-        config[zigpy.config.CONF_DATABASE] = str(data_folder / ("zigpy_persistent_%02d.db"% self.hardwareid) )
-        config[zigpy.config.CONF_TOPO_SCAN_ENABLED] = bool( self.pluginconf.pluginConf["ZigpyAutoTopology"])
-        config[zigpy.config.CONF_TOPO_SCAN_PERIOD] = 12 * 60  # 12 Hours
-
-    elif self.pluginconf.pluginConf.get("enableZigpyPersistentInMemory"):
-        config[zigpy.config.CONF_DATABASE] = ":memory:"
-        config[zigpy.config.CONF_TOPO_SCAN_ENABLED] = bool( self.pluginconf.pluginConf["ZigpyAutoTopology"])
-        config[zigpy.config.CONF_TOPO_SCAN_PERIOD] = 12 * 60  # 12 Hours
-
-    # Manage coordinator auto backup
-    if self.pluginconf.pluginConf.get("autoBackup"):
-        config[zigpy.config.CONF_NWK_BACKUP_ENABLED] = True
-        config[zigpy.config.CONF_NWK_BACKUP_PERIOD] = self.pluginconf.pluginConf["autoBackup"]
-    else:
-        config[zigpy.config.CONF_NWK_BACKUP_ENABLED] = False
-
-    # Do we do energy scan at startup. By default it is set to False. Plugin might override it in the case of low number of devices.
-    if "EnergyScanAtStatup" in self.pluginconf.pluginConf and not self.pluginconf.pluginConf["EnergyScanAtStatup"]:
-        config[zigpy.config.CONF_STARTUP_ENERGY_SCAN] = False
-
-
-async def _radio_startup(self, statistics, pluginconf, use_of_zigpy_persistent_db, new_network, radiomodule):
-    """
-    Performs startup operations for the radio after configuration.
-
-    Starts the application, forms a new network if requested, displays network
-    info, sets the network key, and runs post-startup actions.
-
-    Args:
-        statistics: Statistics tracker.
-        pluginconf: Plugin configuration.
-        use_of_zigpy_persistent_db: Persistent DB flag.
-        new_network (bool): Whether to form a new network.
-        radiomodule (str): Radio module type.
-    """
-    if not self.app:
-        self.log.logging( "TransportZigpy", "Error", "Error at startup - self.app not initialized")
-        return
-    
-    try:
-        await self.app.startup(
-            self.statistics,
-            self.hardwareid,
-            pluginconf,
-            use_of_zigpy_persistent_db,
-            callBackHandleMessage=self.receiveData,
-            callBackUpdDevice=self.ZigpyUpdDevice,
-            callBackGetDevice=self.ZigpyGetDevice,
-            callBackBackup=self.ZigpyBackupAvailable,
-            callBackRestartPlugin=self.restart_plugin,
-            captureRxFrame=self.captureRxFrame,
-            auto_form=True,
-            force_form=new_network,
-            log=self.log,
-            permit_to_join_timer=self.permit_to_join_timer,
-        )
-    except Exception as e:
-        self.log.logging( "TransportZigpy", "Error", "Error at startup %s" %e)
-        return
-
-    if new_network:
-        # Assume that the new network has been created
-        self.log.logging( "TransportZigpy", "Status", "++ Assuming new network formed")
-        self.ErasePDMDone = True  
-
-    display_network_infos(self)
-    self.ControllerData["Network key"] = ":".join( f"{c:02x}" for c in self.app.state.network_information.network_key.key )
-    
-    post_coordinator_startup(self, radiomodule)
-    # Signal to AppGeneric that the supervisor owns this app instance
-    if self.app:
-        self.app._supervisor_running = True
-        self.app.zigpy_running_ref = self
-        self.log.logging(
-            "TransportZigpyStack", "Debug",
-            f"_radio_startup: supervisor refs injected onto app "
-            f"(_supervisor_running=True, zigpy_running_ref={id(self):#x})"
-        )
-
-    
-
-def post_coordinator_startup(self, radiomodule):
-    """
-    Performs actions after coordinator startup, such as sending network info to the plugin.
-
-    Forwards network information, controller endpoints, node descriptors, and
-    simulates an off/on event via plugin frames.
-
-    Args:
-        self: Instance of the transport class.
-        radiomodule (str): Radio module type.
-    """
-    # Send Network information to plugin, in order to poplulate various objetcs
-    self.forwarder_queue.put(build_plugin_8009_frame_content(self, radiomodule))
-
-    # Send Controller Active Node and Node Descriptor
-    self.forwarder_queue.put( build_plugin_8045_frame_list_controller_ep( self, ) )
-
-    self.log.logging( "TransportZigpy", "Debug", "Active Endpoint List:  %s" % str(self.app.get_device(nwk=t.NWK(0x0000)).endpoints.keys()), )
-    for epid, ep in self.app.get_device(nwk=t.NWK(0x0000)).endpoints.items():
-        if epid != 0 and ep.status == 0x00:
-            self.log.logging( "TransportZigpy", "Debug", "Simple Descriptor:  %s" % ep)
-            self.forwarder_queue.put(build_plugin_8043_frame_list_node_descriptor(self, epid, ep))
-
-    self.log.logging( "TransportZigpy", "Debug", "Controller Model %s" % self.app.get_device(nwk=t.NWK(0x0000)).model )
-    self.log.logging( "TransportZigpy", "Debug", "Controller Manufacturer %s" % self.app.get_device(nwk=t.NWK(0x0000)).manufacturer )
-    # Let send a 0302 to simulate an Off/on
-    self.forwarder_queue.put( build_plugin_0302_frame_content( self, ) )
-
-
-def display_network_infos(self):
-    """
-    Logs detailed network information from the Zigpy application state.
-
-    Includes PAN ID, extended PAN ID, channel, channel mask, NWK update ID,
-    device IEEE and NWK, network key, sequence, and counter.
-    """
-    self.log.logging( "TransportZigpy", "Status", "++ Network settings")
-    self.log.logging( "TransportZigpy", "Status", f"  PAN ID:                0x{self.app.state.network_info.pan_id:04X}")
-    self.log.logging( "TransportZigpy", "Status", f"  Extended PAN ID:       {self.app.state.network_info.extended_pan_id}")
-    self.log.logging( "TransportZigpy", "Status", f"  Channel:               {self.app.state.network_info.channel}")
-    self.log.logging( "TransportZigpy", "Status", f"  Channel mask:          {list(self.app.state.network_info.channel_mask)}")
-    self.log.logging( "TransportZigpy", "Status", f"  NWK update ID:         {self.app.state.network_info.nwk_update_id}")
-    self.log.logging( "TransportZigpy", "Status", f"  Device IEEE:           {self.app.state.node_info.ieee}")
-    self.log.logging( "TransportZigpy", "Status", f"  Device NWK:            0x{self.app.state.node_info.nwk:04X}")
-
-    self.log.logging( "TransportZigpy", "Status", "  Network key:           " + ":".join( f"{c:02x}" for c in self.app.state.network_information.network_key.key ))
-    self.log.logging( "TransportZigpy", "Status", f"  Network key sequence:  {self.app.state.network_info.network_key.seq}")
-    self.log.logging( "TransportZigpy", "Status", f"  Network key counter:   {self.app.state.network_info.network_key.tx_counter}")
-
-    self.log.logging( "TransportZigpy", "Status", f"  TX Power:             {self.app.state.network_information.tx_power}")
-    self.log.logging( "TransportZigpy", "Status", f"  Security Level:       {self.app.state.network_information.security_level}")
-    self.log.logging( "TransportZigpy", "Status", f"  Children:             {self.app.state.network_information.children}")
-    self.log.logging( "TransportZigpy", "Status", f"  Routers:              {self.app.state.network_information.route_table}")
-
-                  
-async def worker_loop(self):
-    """
-    Main worker loop for processing commands from the writer_queue.
-
-    Runs while zigpy_running is True, fetches commands, dispatches them,
-    and handles exceptions. Exits on "STOP" command or cancellation.
-    """
-    self.log.logging("TransportZigpy", "Debug", "worker_loop - ZigyTransport: worker_loop start.")
-
-    try:
-        while self.zigpy_running:
-            try:
-                # Fetch the next command to process, waits if queue is empty
-                command_to_send = await get_next_command(self)
-
-                # Break the loop if no command was retrieved due to an error
-                if command_to_send is None:
-                    continue
-
-                # Handle the stop command and exit the loop gracefully
-                if command_to_send == "STOP" or not self.zigpy_running:
-                    self.log.logging(["TransportZigpy", "StopProcess"], "Debug", "worker_loop - Shutting down ... exit.")
-                    self.zigpy_running = False
-                    break
-
-                # Process the received command
-                await process_incoming_command(self, command_to_send)
-
-            except asyncio.CancelledError:
-                # Gracefully handle cancellation
-                self.log.logging("TransportZigpy", "Debug", "worker_loop - Task was cancelled.")
-                break
-
-            except Exception as e:
-                # Log unexpected exceptions but continue processing other commands
-                self.log.logging("TransportZigpy", "Error", f"Unexpected error in worker_loop: {e}")
-
-    finally:
-        # Final cleanup if needed
-        self.log.logging(["TransportZigpy", "StopProcess"], "Debug", "TransportZigpy - Exiting loop and cleaning up resources.")
-
-
-async def process_incoming_command(self, command_to_send):
-    """
-    Processes a single incoming command from the queue.
-
-    Parses JSON data and dispatches to the appropriate handler, catching and
-    logging specific Zigbee-related exceptions.
-
-    Args:
-        self: Instance of the transport class.
-        command_to_send (str): JSON string containing the command data.
-    """
-    data = json.loads(command_to_send)
-    try:
-        await dispatch_command(self, data)
-
-    except (DeliveryError, APIException, ControllerException, InvalidFrame, 
-            CommandNotRecognized, ValueError, InvalidResponse, 
-            InvalidCommandResponse, asyncio.TimeoutError, RuntimeError) as e:
-        log_exception(self, type(e).__name__, e, data.get("cmd", ""), data.get("datas", ""))
-        if isinstance(e, (APIException, ControllerException)):
-            await asyncio.sleep(1.0)
-
-    except Exception as e:
-        self.log.logging("TransportZigpy", "Error", f"Error while receiving a Plugin command: >{e}<")
-        handle_thread_error(self, e, data)
-
-
-async def get_next_command(self):
-    """
-    Asynchronously retrieves the next command from the writer_queue.
-
-    Polls the queue with a short sleep on empty, returns None on error.
-
-    Returns:
-        str or None: The next command string or None on error.
-    """
-    while True:
-        try:
-            return self.writer_queue.get_nowait()
-
-        except queue.Empty:
-            # In case self.zigpy_running is False, we shouldn't wait for a QUIT message
-            # but simply return None, as in worker_loop we will stop iterating 
-            # as self.zigpy_running is False
-            if not self.zigpy_running:
-                return None
-
-            # Otherwise wait a little
-            await asyncio.sleep(0.100)
-
-        except Exception as e:
-            self.log.logging("TransportZigpy", "Log", f"Error in get_next_command: {e}")
-            return None
-
-
-async def dispatch_command(self, data):
-    """
-    Dispatches the parsed command data to the appropriate handler function.
-
-    Supports commands like backup, permit-to-join, raw commands, device removal,
-    network status, certification, channel/PAN ID/LED/TX power/time setting,
-    and scans.
-
-    Args:
-        self: Instance of the transport class.
-        data (dict): Parsed JSON command data with 'cmd' and 'datas' keys.
-    """
-    cmd = data["cmd"]
-    datas = data["datas"]
-    delayAfterSent = datas.get("delayAfterSent", 0) if datas else 0
-
-    if cmd == "COORDINATOR-BACKUP":
-        await self.app.coordinator_backup()
-
-    elif cmd == "GET-TIME":
-        await self.app.get_time_server()
-
-    elif cmd == "PERMIT-TO-JOIN":
-        await _permit_to_joint(self, data)
-
-    elif cmd == "RAW-COMMAND":
-        self.log.logging("TransportZigpy", "Debug", f"RAW-COMMAND: {properyly_display_data(datas)}")
-        await process_raw_command(self, datas, AckIsDisable=data["ACKIsDisable"], Sqn=data["Sqn"], delayAfterSent=delayAfterSent)
-
-    elif cmd == "REMOVE-DEVICE":
-        ieee = datas["Param1"]
-        await self.app.remove_ieee(t.EUI64(t.uint64_t(ieee).serialize()))
-
-    elif cmd == "REQ-NWK-STATUS":
-        await asyncio.sleep(10)
-        self.forwarder_queue.put(build_plugin_8009_frame_content(self, self._radiomodule))
-
-    elif cmd == "SET-CERTIFICATION":
-        await self.app.set_certification(datas["Param1"])
-
-    elif cmd == "SET-CHANNEL":
-        await self.app.move_network_to_channel(datas["Param1"])
-
-    elif cmd == "SET-EXTPANID":
-        self.app.set_extended_pan_id(datas["Param1"])
-
-    elif cmd == "SET-LED":
-        await self.app.set_led(datas["Param1"])
-
-    elif cmd == "SET-TIME":
-        await self.app.set_time_server(datas["Param1"])
-
-    elif cmd == "SET-TX-POWER":
-        await self.app.set_zigpy_tx_power(datas["Param1"])
-        
-    elif cmd == "INTERFERENCE-SCAN":
-        self.manual_interference_scan_task = asyncio.create_task( self.app.network_interference_scan(), name="INTERFERENCE-SCAN")
-
-    elif cmd == "ZIGPY-TOPOLOGY-SCAN":
-        self.manual_topology_scan_task = asyncio.create_task( self.app.start_topology_scan(), name="ZIGPY-TOPOLOGY-SCAN")
-
-    elif cmd == "RESTART-ZIGPY-STACK":
-        # Graceful stack restart: exit worker_loop → start_zigpy_task exits →
-        # supervisor restarts the full stack.  Does NOT restart the Domoticz plugin.
-        self.log.logging("TransportZigpy", "Log",
-                         "RESTART-ZIGPY-STACK: graceful stack restart requested")
-        self.zigpy_running = False
-        self.writer_queue.put_nowait("STOP")
-
-    elif cmd == "RESET-RADIO-COMMUNICATION":
-        # Soft reset: disconnect and reconnect the transport layer only.
-        # The zigpy stack (network state, device table) is preserved.
-        # Falls back to a full stack restart if reconnect fails.
-        self.log.logging("TransportZigpy", "Log",
-                         "RESET-RADIO-COMMUNICATION: transport reconnect requested")
-        if self.app:
-            try:
-                await asyncio.wait_for(self.app.disconnect(), timeout=10.0)
-                await asyncio.sleep(2)
-                await asyncio.wait_for(self.app.connect(), timeout=15.0)
-                self.log.logging("TransportZigpy", "Status",
-                                 "RESET-RADIO-COMMUNICATION: radio communication reset complete")
-            except Exception as e:
-                self.log.logging("TransportZigpy", "Error",
-                                 f"RESET-RADIO-COMMUNICATION failed ({e}) — falling back to full stack restart")
-                self.zigpy_running = False
-                self.writer_queue.put_nowait("STOP")
-
-
-async def _permit_to_joint(self, data):
-    """
-    Handles the PERMIT-TO-JOIN command to open the network for device joining.
-
-    Sets a timer and calls the app's permit method, with special handling for deCONZ.
-
-    Args:
-        self: Instance of the transport class.
-        data (dict): Command data with 'datas' containing Duration and targetRouter.
-    """
-    log = self.log
-    radiomodule = self._radiomodule
-    app = self.app
-    permit_to_join_timer = self.permit_to_join_timer
-
-    log.logging("TransportZigpy", "Debug", f"PERMIT-TO-JOIN: {data}")
-
-    duration = data["datas"]["Duration"]
-    target_router = data["datas"]["targetRouter"]
-    target_router = None if target_router == "FFFC" else t.EUI64(t.uint64_t(target_router).serialize())
-    duration = 0xFE if duration == 0xFF else duration
-
-    permit_to_join_timer["Timer"] = time.time()
-    permit_to_join_timer["Duration"] = duration
-
-    log.logging("TransportZigpy", "Status", f"++ opening zigbee network for {duration} secondes on specific router {target_router}")
-
-    if radiomodule == "deCONZ":
-        return await app.permit_ncp(time_s=duration)
-
-    log.logging("TransportZigpy", "Debug", f"Calling app.permit(time_s={duration}, node={target_router})")
-    await app.permit(time_s=duration, node=target_router)
-    log.logging("TransportZigpy", "Debug", f"Returning from app.permit(time_s={duration}, node={target_router})")
-
-
-async def process_raw_command(self, data, AckIsDisable=False, Sqn=None, delayAfterSent=0):
-    """
-    Processes a raw Zigbee command and determines the transmission type.
-
-    Extracts parameters, determines destination and transport type (broadcast,
-    multicast, unicast), and calls the appropriate send function.
-
-    Args:
-        self: Instance of the transport class.
-        data (dict): Raw command data with keys like Function, Profile, Cluster, etc.
-        AckIsDisable (bool): Whether to disable APS ACK.
-        Sqn (int): Sequence number.
-        delayAfterSent (float): Delay after sending.
-
-    Returns:
-        None
-    """
-    Function = data["Function"]
-    TimeStamp = data["timestamp"]
-    Profile = data["Profile"]
-    Cluster = data["Cluster"]
-    NwkId = "%04x" % data["TargetNwk"]
-    dEp = data["TargetEp"]
-    sEp = data["SrcEp"]
-    payload = bytes.fromhex(data["payload"])
-    sequence = Sqn or self.app.get_sequence()
-    addressmode = data["AddressMode"]
-    extended_timeout = False if AckIsDisable else data.get("RxOnIdle", False)
-    delay = data.get("Delay", None)
-
-    self.log.logging("TransportZigpy", "Debug", f"process_raw_command: ready to request Function: {Function} NwkId: {NwkId}/{dEp} Cluster: {Cluster} Seq: {sequence} Payload: {payload.hex()} AddrMode: {addressmode} AckIsDisable: {AckIsDisable} Sqn: {Sqn}, Delay: {delay}, delayAfterSent {delayAfterSent}, Extended_TO: {extended_timeout}")
-
-    destination, transport_needs = _get_destination(self, NwkId, addressmode, Profile, Cluster, sEp, dEp, sequence, payload)
-    if destination is None:
-        self.log.logging("TransportZigpy", "Error", f"process_raw_command: unknown destination: {destination}")
-        return
-
-    handlers = {
-        "Broadcast": lambda: send_broadcast_command(self, Profile, Cluster, sEp, dEp, sequence, payload),
-        "Multicast": lambda: send_multicast_command(self, NwkId, Profile, Cluster, sEp, sequence, payload),
-        "Unicast": lambda: send_unicast_command(self, destination, Profile, Cluster, sEp, dEp, sequence, payload, AckIsDisable, delay, extended_timeout, Function, Sqn, delayAfterSent),
-    }
-
-    key = "Multicast" if addressmode == 0x01 else transport_needs
-    handler = handlers.get(key)
-    if handler is None:
-        self.log.logging("TransportZigpy", "Error", f"process_raw_command: unhandled transport '{transport_needs}' addrmode {addressmode}")
-        return
-
-    result, msg = await handler()
-    self.log.logging("TransportZigpy", "Debug", f"process_raw_command completed: {destination} result={result} msg={msg}")
-
-
-async def send_broadcast_command(self, Profile, Cluster, sEp, dEp, sequence, payload):
-    """
-    Sends a broadcast Zigbee command.
-
-    Uses app.broadcast and adds a wait between requests.
-
-    Args:
-        self: Instance of the transport class.
-        Profile (int): Zigbee profile ID.
-        Cluster (int): Zigbee cluster ID.
-        sEp (int): Source endpoint.
-        dEp (int): Destination endpoint.
-        sequence (int): Sequence number.
-        payload (bytes): Command payload.
-
-    Returns:
-        tuple: (result, message) from the broadcast operation.
-    """
-    result, msg = await self.app.broadcast(Profile, Cluster, sEp, dEp, 0x0, 0x0, sequence, payload)
-    await asyncio.sleep(2 * WAITING_TIME_BETWEEN_REQUESTS)
-    return result, msg
-
-
-async def send_multicast_command(self, NwkId, Profile, Cluster, sEp, sequence, payload):
-    """
-    Sends a multicast Zigbee command to a group.
-
-    Uses app.mrequest with the group ID.
-
-    Args:
-        self: Instance of the transport class.
-        NwkId (str): Group ID as hex string.
-        Profile (int): Zigbee profile ID.
-        Cluster (int): Zigbee cluster ID.
-        sEp (int): Source endpoint.
-        sequence (int): Sequence number.
-        payload (bytes): Command payload.
-
-    Returns:
-        tuple: (result, message) from the multicast operation.
-    """
-    destination = int(NwkId, 16)
-    self.log.logging("TransportZigpy", "Debug", f"send_multicast_command Multicast: {destination}")
-    result, msg = await self.app.mrequest(destination, Profile, Cluster, sEp, sequence, payload)
-    await asyncio.sleep(2 * WAITING_TIME_BETWEEN_REQUESTS)
-    return result, msg
-
-
-async def send_unicast_command(self, destination, Profile, Cluster, sEp, dEp, sequence, payload, AckIsDisable, delay, extended_timeout, Function, Sqn, delayAfterSent):
-    """
-    Sends a unicast command to a Zigbee device.
-
-    Args:
-        destination (str): The destination address of the Zigbee device.
-        Profile (int): The Zigbee profile ID.
-        Cluster (int): The Zigbee cluster ID.
-        sEp (int): Source endpoint.
-        dEp (int): Destination endpoint.
-        sequence (int): Sequence number for the command.
-        payload (bytes): The command payload.
-        AckIsDisable (bool): Whether to disable APS acknowledgments.
-        delay (float): Delay before sending the command.
-        extended_timeout (float): Timeout for the task.
-        Function (str): Function identifier for the command.
-        Sqn (int): Sequence number for task naming.
-        delayAfterSent (float): Delay after sending the command.
-
-    Returns:
-        tuple: (result, error_msg)
-            - result (int): Status code (0x00 for success, error code for failure).
-            - error_msg (str): Error message if the task creation fails.
-    """
-
-    payload_hex = payload.hex()[:100] + "..." if len(payload.hex()) > 100 else payload.hex()
-    AckIsDisable = False if self.pluginconf.pluginConf["ForceAPSAck"] else AckIsDisable
-
-    self.log.logging("TransportZigpy", "Debug", f"send_unicast_command Unicast destination: {destination} Profile: {Profile} Cluster: {Cluster} sEp: {sEp} dEp: {dEp} Seq: {sequence} Payload: {payload_hex}")
-
-    try:
-        task = asyncio.create_task(
-            transport_request(self, Function, destination, Profile, Cluster, sEp, dEp, sequence, payload, ack_is_disable=AckIsDisable, use_ieee=False, delay=delay, extended_timeout=extended_timeout, delayAfterSent=delayAfterSent),
-            name=f"send_unicast_command-{Function}-{destination}-{Cluster}-{Sqn}"
-        )
-
-    except (TypeError, ValueError, RuntimeError) as e:
-        self.log.logging("TransportZigpy", "Error", f"Failed to create task: {e}")
-        self.statistics._ackKO += 1
-        return ERROR_TASK_CREATION_FAILED, str(e)
-
-    task.add_done_callback(_make_unicast_callback(self))
-    self.statistics._sent += 1
-    return 0x00, ""
-
-
-def _make_unicast_callback(self):
-    """Returns a task-done callback that logs and updates statistics."""
-
-    def callback(task):
-        if task.cancelled():
-            return
-        exc = task.exception()
-        if exc:
-            self.statistics._ackKO += 1
-            self.log.logging("TransportZigpy", "Debug", f"Task {task.get_name()} failed: {exc}")
-        else:
-            self.log.logging("TransportZigpy", "Debug", f"Task {task.get_name()} completed")
-    return callback
-
-
-def _get_destination(self, NwkId, addressmode, Profile, Cluster, sEp, dEp, sequence, payload):
-    """
-    Determines the destination device and transport type for a command.
-
-    Handles broadcast, multicast, and unicast based on address mode and NWK ID.
-
-    Args:
-        self: Instance of the transport class.
-        NwkId (str): Network ID as hex string.
-        addressmode (int): Addressing mode (0x00-0x08).
-        Profile (int): Profile ID.
-        Cluster (int): Cluster ID.
-        sEp (int): Source endpoint.
-        dEp (int): Destination endpoint.
-        sequence (int): Sequence number.
-        payload (bytes): Payload.
-
-    Returns:
-        tuple: (destination object or int, str transport type) or (None, None) on error.
-    """
-    if int(NwkId, 16) >= 0xFFFB:  
-        # Broadcast
-        return int(NwkId, 16), "Broadcast"
-
-    if addressmode == 0x01:
-        # Group
-        return int(NwkId, 16), "Multicast"
-
-    if addressmode in (0x02, 0x07):
-        # 0x02 Short address
-        # 0x07 Short address with No Ack (Zigate)
-        try:
-            destination = self.app.get_device(nwk=t.NWK(int(NwkId, 16)))
-
-        except KeyError:
-            self.log.logging( "TransportZigpy", "Log", f"_get_destination unable to get destination. Nwkid {NwkId} AddrMode {addressmode}")
-            destination = None
-
-        return destination, "Unicast"
-
-    if addressmode in (0x03, 0x08):
-        # 0x03 IEEE
-        # 0x08 IEEE with No Ack (Zigate)
-        return self.app.get_device(nwk=t.NWK(int(NwkId, 16))), "Unicast"
-    
-    self.log.logging( "TransportZigpy", "Error", f"_get_destination wrong address mode {addressmode} NwkId {NwkId}")
-    return None, None
-
-
-def push_APS_ACK_NACKto_plugin(self, nwkid, Cluster, sequence, result, lqi):
-    """
-    Forwards APS ACK/NACK status to the plugin via forwarder_queue.
-
-    Updates statistics and skips for coordinator (nwkid=0000). Converts result
-    to int if necessary.
-
-    Args:
-        self: Instance of the transport class.
-        nwkid (str): Network ID as hex.
-        Cluster (int): Cluster ID.
-        sequence (int): Sequence number.
-        result: Result status (int or serializable).
-        lqi (int): Link Quality Indicator.
-    """
-    # Looks like Zigate return an int, while ZNP returns a status.type
-    self.log.logging("TransportZigpy", "Debug", f"push_APS_ACK_NACK to_plugin - {nwkid} - Result: {result} LQI: {lqi}")
-    if nwkid == "0000":
-        # No Ack/Nack for Controller
-        return
-    
-    try:
-        if not isinstance(result, int):
-            result = int(result.serialize().hex(), 16)
-        # Update statistics
-        if result != 0x00:
-            self.statistics._APSNck += 1
-        else:
-            self.statistics._APSAck += 1
-
-    except Exception as e:
-        result = -1
-        self.statistics._APSNck += 1
-
-    # Send Ack/Nack to Plugin
-    self.forwarder_queue.put(build_plugin_8011_frame_content(self, nwkid, Cluster, sequence, result, lqi))
-
-
-def properyly_display_data(Datas):
-    """
-    Formats a dictionary of data into a readable log string.
-
-    Converts specific keys (Profile, Cluster, etc.) to hex format for display.
-
-    Args:
-        Datas (dict): Data dictionary to format.
-
-    Returns:
-        str: Formatted log string like "{'key': value, ...}".
-    """
-    log = "{"
-    for x in Datas:
-        value = Datas[x]
-        if x in (
-            "Profile",
-            "Cluster",
-            "TargetNwk",
-        ):
-            if isinstance(value, int):
-                value = "%04x" % value
-        elif x in ("TargetEp", "SrcEp", "Sqn", "AddressMode"):
-            if isinstance(value, int):
-                value = "%02x" % value
-        log += "'%s' : %s," % (x, value)
-    log += "}"
-    return log
-
-
-def log_exception(self, exception, error, cmd, data):
-    """
-    Logs an exception with context including stack trace and command data.
-
-    Uses properyly_display_data for formatting data.
-
-    Args:
-        self: Instance of the transport class.
-        exception (str): Exception type name.
-        error: Exception instance.
-        cmd (str): Command name.
-        data (dict): Command data.
-    """
-    context = {
-        "Exception": str(exception),
-        "Message code:": str(error),
-        "Stack Trace": str(traceback.format_exc()),
-        "Command": str(cmd),
-        "Data": properyly_display_data(data),
-    }
-
-    self.log.logging(
-        "TransportZigpy",
-        "Error",
-        "%s / %s: request() Not able to execute the zigpy command: %s data: %s"
-        % (exception, error, cmd, properyly_display_data(data)),
-        context=context,
-    )
-
-
-def check_transport_readiness(self):
-    """
-    Checks if the transport is ready based on the radio module.
-
-    Returns True for zigate/deCONZ/ezsp/blz, and checks ZNP sequence for znp.
-
-    Returns:
-        bool: True if ready, False otherwise.
-    """
-    radiomodule = self._radiomodule
-    if radiomodule in {"zigate", "deCONZ", "ezsp", "blz"}:
-        return True
-
-    if radiomodule == "znp":
-        app = self.app
-        return app._znp is not None
-
-    return False
-
-
-def measure_execution_time(func):
-    """
-    Decorator to measure and log execution time of async functions.
-
-    Logs timing if ZigpyReactTime is enabled, updates statistics, and logs
-    detailed device info on completion.
-
-    Args:
-        func (callable): Async function to decorate.
-
-    Returns:
-        callable: Wrapped function.
-    """
-    @functools.wraps(func)
-    async def wrapper(self, *args, **kwargs):
-        t_start = None
-        if getattr(self, "pluginconf", None) and self.pluginconf.pluginConf.get("ZigpyReactTime", False):
-            t_start = time.time()
-
-        try:
-            result = await func(self, *args, **kwargs)
-            return result
-        finally:
-            if t_start:
-                t_end = time.time()
-                t_elapse = round((t_end - t_start) * 1000)  # milliseconds
-
-                if hasattr(self, "statistics"):
-                    self.statistics.add_timing_zigpy(t_elapse)
-
-                if hasattr(self, "log"):
-                    # Safely extract info from known kwargs or fallback
-                    Function = kwargs.get("Function", args[0] if len(args) > 0 else "Unknown")
-                    sequence = kwargs.get("sequence", args[6] if len(args) > 6 else "N/A")
-                    ack_is_disable = kwargs.get("ack_is_disable", args[7] if len(args) > 7 else False)
-                    destination = kwargs.get("destination", args[1] if len(args) > 1 else None)
-
-                    nwk = getattr(destination.nwk, "hex", lambda: "??")() if destination else "??"
-                    ieee = getattr(destination, "ieee", "??")
-                    model = getattr(destination, "model", "??")
-                    mfr = getattr(destination, "manufacturer_id", "??")
-                    init = getattr(destination, "is_initialized", "??")
-                    rssi = getattr(destination, "rssi", "??")
-                    lqi = getattr(destination, "lqi", "??")
-
-                    self.log.logging(
-                        "TransportZigpy", "Log",
-                        f"| (transport_request) | {t_elapse}ms | {Function} | {sequence} | {ack_is_disable} | {nwk} | {ieee} | {model} | {mfr} | {init} | {rssi} | {lqi} |"
-                    )
-    return wrapper
-
-
-@measure_execution_time
-async def transport_request(
-    self,
-    Function,
-    destination,
-    Profile,
-    Cluster,
-    sEp,
-    dEp,
-    sequence,
-    payload,
-    ack_is_disable=False,
-    use_ieee=False,
-    delay=None,
-    extended_timeout=False,
-    delayAfterSent=0
-):
-    """
-    Send a Zigbee message using the transport layer.
-
-    Args:
-        Function (str): Operation name for logging and stats.
-        destination: Zigpy device object (must have `nwk` and `ieee` attributes).
-        Profile (int): Zigbee Profile ID (e.g., 0x0000 or 0x0104).
-        Cluster (int): Zigbee Cluster ID.
-        sEp (int): Source endpoint.
-        dEp (int): Destination endpoint.
-        sequence (int): Transaction sequence number.
-        payload (bytes): Zigbee payload data.
-        ack_is_disable (bool, optional): If True, disables waiting for ACK. Defaults to False.
-        use_ieee (bool, optional): If True, uses IEEE addressing. Defaults to False.
-        delay (float, optional): Optional delay (in seconds) before sending. Defaults to None.
-        extended_timeout (bool, optional): Enables extended timeout. Defaults to False.
-        delayAfterSent (float, optional): Delay after sending the request. Defaults to 0.
-
-    Returns:
-        int | None: Zigbee transmission result code (e.g., 0x00 for success, 0xB6 for error), or None if skipped.
-    """
-
-    _nwkid = destination.nwk.serialize()[::-1].hex()
-    _ieee = str(destination.ieee)
-
-    if not check_transport_readiness(self):
-        return None
-
-    # Optional delay for specific devices (e.g., CASA.IA)
-    if Profile == 0x0000 and Cluster == 0x0005 and _ieee and _ieee[:8] in DELAY_FOR_VERY_KEY:
-        self.log.logging("TransportZigpy", "Debug", f"Delaying for key verification for {_ieee}")
-        delay = delay or VERIFY_KEY_DELAY  # Let VERIFY_KEY_DELAY be something like 1.0 instead of 6.0 if desired
-
-    if delay:
-        self.log.logging("TransportZigpy", "Debug", f"transport_request: delay for {delay} seconds")
-        await asyncio.sleep(delay)
-
-    async with _limit_concurrency(self, destination, sequence):
-        if _ieee in self._currently_not_reachable and self._currently_waiting_requests_list.get(_ieee, 0):
-            self.log.logging(
-                "TransportZigpy", "Debug",
-                f"transport_request: Request {sequence} skipped. Device not reachable: NwkId: {_nwkid} IEEE: {_ieee}"
-            )
-            return None
-
-        result = await _send_and_retry(
-            self,
-            Function,
-            destination,
-            Profile,
-            Cluster,
-            _nwkid,
-            sEp,
-            dEp,
-            sequence,
-            payload,
-            use_ieee,
-            _ieee,
-            ack_is_disable,
-            extended_timeout,
-            delayAfterSent
-        )
-
-    await asyncio.sleep(WAITING_TIME_BETWEEN_REQUESTS)
-    return result
-
-
-async def _send_and_retry(
-    self, function, destination, profile, cluster,
-    nwkid, source_ep, dest_ep, sequence, payload,
-    use_ieee, ieee, ack_is_disable, extended_timeout, delay_after_sent
-):
-    """
-    Sends a Zigbee request and optionally retries the operation until a timeout is reached.
-
-    If `ack_is_disable` is False, the function attempts to send the request only once.
-    If `ack_is_disable` is True, the function retries the request until either a valid response
-    is received or `REQUEST_TIMEOUT` is exceeded.
-
-    Parameters:
-        function (str): Description of the operation being performed (used for logging).
-        destination: Zigbee destination object.
-        profile (int): Zigbee profile ID.
-        cluster (int): Zigbee cluster ID.
-        nwkid (int): Network address of the destination.
-        source_ep (int): Source endpoint.
-        dest_ep (int): Destination endpoint.
-        sequence (int): Sequence number for the request.
-        payload (bytes): Payload to send.
-        use_ieee (bool): Whether to address the device using IEEE address.
-        ieee (str): IEEE address of the device (used for logging).
-        ack_is_disable (bool): If True, ACK is disabled and retry logic is enabled.
-        extended_timeout (bool): Whether to use an extended timeout for the Zigbee request.
-        delay_after_sent (float): Optional delay after a successful transmission.
-
-    Returns:
-        int: The Zigbee result code (e.g., 0x00 for success, 0xB6 for timeout or fatal error).
-    """
-    common_log_info = f"{ieee}/0x{nwkid} 0x{profile:X} 0x{cluster:X} payload: {payload} AckIsDisable: {ack_is_disable} extended_timeout: {extended_timeout}"
-    packet_priority = t.PacketPriority.NORMAL
-
-    async def __try_send(attempt):
-        """
-        Sends a Zigbee request and handles exceptions.
-
-        Returns:
-            int | None: Result code if successful or handled exception; None if retry is needed.
-        """
-        self.log.logging("TransportZigpy", "Debug",
-                         f"_send_and_retry: {function} {common_log_info} "
-                         f"extended_timeout: {extended_timeout} Attempt: {attempt}")
-
-        try:
-            result, _ = await zigpy_request(
-                self, destination, profile, cluster,
-                source_ep, dest_ep, sequence, payload,
-                ack_is_disable=ack_is_disable,
-                use_ieee=use_ieee,
-                extended_timeout=extended_timeout,
-                priority=packet_priority
-            )
-
-        except asyncio.TimeoutError:
-            msg = f"Timeout while submitting - {function} {common_log_info} Attempt: {attempt}"
-            self.log.logging("TransportZigpy", "Debug", msg)
-            self.statistics._reTx += 1
-            self.statistics._TOdata += 1
-
-            # Let's retry and so return None
-            return None
-
-        except Exception as e:
-            msg = f"Warning while submitting - {function} {common_log_info} Attempt: {attempt} Exception: '{e}' ({type(e).__name__})"
-            self.log.logging("TransportZigpy", "Log", msg)
-            self.statistics._ackKO += 1
-
-            # No retry, return 0xB6
-            handle_transport_result(self, function, cluster, sequence, 0xB6, ack_is_disable, extended_timeout, ieee, nwkid, destination.lqi)
-            return 0xB6
-
-        else:
-            if result == -1:
-                # UART Closed
-                return result
-
-            delay_after_cmd = max(delay_after_sent, self.pluginconf.pluginConf.get("DelayAfterCommandSent", 0))
-            if delay_after_cmd > 0:
-                self.log.logging("TransportZigpy", "Debug", f"sleeping {delay_after_cmd} as per configured!!")
-                await asyncio.sleep(delay_after_cmd)
-
-            # Successful transmission
-            handle_transport_result(self, function, cluster, sequence, result, ack_is_disable, extended_timeout, ieee, nwkid, destination.lqi)
-            self.log.logging("TransportZigpy", "Debug", f"_send_and_retry: result: {result}")
-            return result
-   
-    if ack_is_disable:
-        # Single send, no retry
-        return await __try_send(attempt=1)
-
-    # Retry mode for ack_is_disable == False
-    start_time = time.monotonic()
-    attempt = 0
-
-    while True:
-        attempt += 1
-        elapsed = time.monotonic() - start_time
-
-        if elapsed >= REQUEST_TIMEOUT:
-            self.log.logging("TransportZigpy", "Log", f"_send_and_retry: {common_log_info} TIMEOUT of {REQUEST_TIMEOUT}s reached after {attempt - 1} attempts. ")
-            self.statistics._ackKO += 1
-            handle_transport_result(self, function, cluster, sequence, 0xB6, ack_is_disable, extended_timeout, ieee, nwkid, destination.lqi)
-            return 0xB6
-
-        self.log.logging("TransportZigpy", "Debug",
-                         f"_send_and_retry: {function} {common_log_info} "
-                         f"extended_timeout: {extended_timeout} Attempt: {attempt} Elapsed: {elapsed:.2f}s")
-        result = await __try_send(attempt)
-        
-        # if result is None it means we need to retry
-        if result is not None:
-            return result
-
-        packet_priority = t.PacketPriority.HIGH  # escalate on retry
-
-
-async def zigpy_request( self, device: zigpy.device.Device, profile: t.uint16_t, cluster: t.uint16_t, src_ep: t.uint8_t, dst_ep: t.uint8_t, sequence: t.uint8_t, data: bytes, *, ack_is_disable: bool = True, use_ieee: bool = False, extended_timeout: bool = False, priority: bool = t.PacketPriority.NORMAL) -> tuple[zigpy.zcl.foundation.Status, str]:
-    """
-    Submits a unicast Zigbee packet via the app.send_packet method.
-
-    Builds the ZigbeePacket with addressing, options, and source routing if enabled.
-    Logs errors on failure.
-
-    Args:
-        self: Instance of the transport class.
-        device: Target Zigpy device.
-        profile: Profile ID.
-        cluster: Cluster ID.
-        src_ep: Source endpoint.
-        dst_ep: Destination endpoint.
-        sequence: TSN sequence.
-        data: Payload bytes.
-        ack_is_disable: Disable ACK (default True).
-        use_ieee: Use IEEE addressing (default False).
-        extended_timeout: Extended timeout flag.
-        priority: Packet priority.
-
-    Returns:
-        tuple: (Status, str message) - SUCCESS on send, DeliveryError on failure.
-    """
-    self.log.logging(
-        "TransportZigpy", 
-        "Debug", 
-        f"zigpy_request: "
-        f"zigpy_request called with: device={device}, profile={profile}, cluster={cluster}, "
-        f"src_ep={src_ep}, dst_ep={dst_ep}, sequence={sequence}, data={data}, "
-        f"ack_is_disable={ack_is_disable}, use_ieee={use_ieee}, extended_timeout={extended_timeout}"
-    )
-    if self.app is None:
-        self.log.logging( "TransportZigpy", "Log", "zigpy_request: app is None, cannot send packet" )
-        return (zigpy.zcl.foundation.Status.DELIVERY_ERROR, "ZCL FAILURE: app is None")
-
-    if use_ieee:
-        src = t.AddrModeAddress( addr_mode=t.AddrMode.IEEE, address=self.app.state.node_info.ieee )
-        dst = t.AddrModeAddress(addr_mode=t.AddrMode.IEEE, address=device.ieee)
-    else:
-        src = t.AddrModeAddress( addr_mode=t.AddrMode.NWK, address=self.app.state.node_info.nwk )
-        dst = t.AddrModeAddress(addr_mode=t.AddrMode.NWK, address=device.nwk)
-
-    if self.app.config[zigpy.config.CONF_SOURCE_ROUTING]:
-        source_route = self.app.build_source_route_to(dest=device)
-    else:
-        source_route = None
-
-    tx_options = t.TransmitOptions.NONE
-
-    if not ack_is_disable:
-        tx_options |= t.TransmitOptions.ACK
-
-    try:
-        await self.app.send_packet(
-            t.ZigbeePacket(
-                src=src,
-                src_ep=src_ep,
-                dst=dst,
-                dst_ep=dst_ep,
-                tsn=sequence,
-                profile_id=profile,
-                cluster_id=cluster,
-                data=t.SerializableBytes(data),
-                extended_timeout=extended_timeout,
-                source_route=source_route,
-                tx_options=tx_options,
-                priority=priority,
-            )
-        )
-    except (AttributeError, asyncio.CancelledError):
-        # Transport already closed or shutdown in progress
-        return -1, None
-
-    except asyncio.TimeoutError as e:
-        self.log.logging(
-            "TransportZigpy",
-            "Debug",
-            (
-                "zigpy_request: Timeout while sending packet\n"
-                f"  src={src}, src_ep={src_ep}, dst={dst}, dst_ep={dst_ep}, tsn={sequence}\n"
-                f"  profile_id={profile}, cluster_id={cluster}, data={data.hex() if isinstance(data,(bytes,bytearray)) else data}\n"
-                f"  extended_timeout={extended_timeout}, source_route={source_route}, "
-                f"tx_options={tx_options}, priority={priority}\n"
-                f"  Exception={e}\n"
-                f"  Traceback:\n{traceback.format_exc()}"
-            ),
-        )
-        return (asyncio.TimeoutError, f"ZCL FAILURE: {e}")
-  
-    except zigpy.exceptions.DeliveryError as e:
-        self.log.logging(
-            "TransportZigpy",
-            "Debug",
-            (
-                "zigpy_request: Error sending packet\n"
-                f"  src={src}, src_ep={src_ep}, dst={dst}, dst_ep={dst_ep}, tsn={sequence}\n"
-                f"  profile_id={profile}, cluster_id={cluster}, data={data.hex() if isinstance(data,(bytes,bytearray)) else data}\n"
-                f"  extended_timeout={extended_timeout}, source_route={source_route}, "
-                f"tx_options={tx_options}, priority={priority}\n"
-                f"  Exception={e}\n"
-                f"  Traceback:\n{traceback.format_exc()}"
-            ),
-        )
-        return (zigpy.exceptions.DeliveryError, f"ZCL FAILURE: {e}")
-      
-    except Exception as e:
-        self.log.logging(
-            "TransportZigpy",
-            "Error",
-            (
-                "zigpy_request: Error sending packet\n"
-                f"  src={src}, src_ep={src_ep}, dst={dst}, dst_ep={dst_ep}, tsn={sequence}\n"
-                f"  profile_id={profile}, cluster_id={cluster}, data={data.hex() if isinstance(data,(bytes,bytearray)) else data}\n"
-                f"  extended_timeout={extended_timeout}, source_route={source_route}, "
-                f"tx_options={tx_options}, priority={priority}\n"
-                f"  Exception={e}\n"
-                f"  Traceback:\n{traceback.format_exc()}"
-            ),
-        )
-        return (zigpy.exceptions.DeliveryError, f"ZCL FAILURE: {e}")
-
-    return (zigpy.zcl.foundation.Status.SUCCESS, "")
-
-
-async def zigpy_mrequest( self, group_id: t.uint16_t, profile: t.uint8_t, cluster: t.uint16_t, src_ep: t.uint8_t, sequence: t.uint8_t, data: bytes, *, hops: int = 0, non_member_radius: int = 3,):
-    """
-    Submits a multicast Zigbee packet to a group.
-
-    Builds and sends the packet with group addressing.
-
-    Args:
-        self: Instance of the transport class.
-        group_id: Group ID.
-        profile: Profile ID.
-        cluster: Cluster ID.
-        src_ep: Source EP.
-        sequence: TSN.
-        data: Payload.
-        hops: Radius for hops.
-        non_member_radius: Non-member radius.
-
-    Returns:
-        tuple: (Status.SUCCESS, "") on send.
-    """
-
-    await self.app.send_packet(
-        t.ZigbeePacket(
-            src=t.AddrModeAddress( addr_mode=t.AddrMode.NWK, address=self.app.state.node_info.nwk ),
-            src_ep=src_ep,
-            dst=t.AddrModeAddress( addr_mode=t.AddrMode.Group, address=group_id),
-            tsn=sequence,
-            profile_id=profile,
-            cluster_id=cluster,
-            data=t.SerializableBytes(data),
-            tx_options=t.TransmitOptions.NONE,
-            radius=hops,
-            non_member_radius=non_member_radius,
-        )
-    )
-
-    return (zigpy.zcl.foundation.Status.SUCCESS, "")
-
-
-async def zigpy_broadcast( self, profile: t.uint16_t, cluster: t.uint16_t, src_ep: t.uint8_t, dst_ep: t.uint8_t, grpid: t.uint16_t, radius: int, sequence: t.uint8_t, data: bytes, broadcast_address: t.BroadcastAddress = t.BroadcastAddress.RX_ON_WHEN_IDLE, ) -> tuple[zigpy.zcl.foundation.Status, str]:
-    """
-    Submits a broadcast Zigbee packet.
-
-    Builds and sends the packet with broadcast addressing.
-
-    Args:
-        self: Instance of the transport class.
-        profile: Profile ID.
-        cluster: Cluster ID.
-        src_ep: Source EP.
-        dst_ep: Dest EP.
-        grpid: Group ID (unused?).
-        radius: Broadcast radius.
-        sequence: TSN.
-        data: Payload.
-        broadcast_address: Broadcast address type.
-
-    Returns:
-        tuple: (Status.SUCCESS, "") on send.
-    """
-    await self.app.send_packet(
-        t.ZigbeePacket(
-            src=t.AddrModeAddress( addr_mode=t.AddrMode.NWK, address=self.app.state.node_info.nwk ),
-            src_ep=src_ep,
-            dst=t.AddrModeAddress( addr_mode=t.AddrMode.Broadcast, address=broadcast_address ),
-            dst_ep=dst_ep,
-            tsn=sequence,
-            profile_id=profile,
-            cluster_id=cluster,
-            data=t.SerializableBytes(data),
-            tx_options=t.TransmitOptions.NONE,
-            radius=radius,
-        )
-    )
-
-    return (zigpy.zcl.foundation.Status.SUCCESS, "")
-
-
-def handle_transport_result(self, Function, Cluster, sequence, result, ack_is_disable, extended_timeout, _ieee, _nwkid, lqi):
-    """
-    Handle the result of a Zigbee transport operation by updating plugin state,
-    acknowledging APS delivery status, and tracking device reachability.
-
-    Args:
-        Function: Zigbee command/function used.
-        Cluster: Zigbee cluster involved in the request.
-        sequence: Sequence number of the request.
-        result (int): APS ACK/NACK result (0x00 = success).
-        ack_is_disable (bool): True if APS ACK was disabled in request.
-        extended_timeout (bool): True if extended timeout was used.
-        _ieee (str): IEEE address of the target device.
-        _nwkid (str): Network ID (NWK address) of the device.
-        lqi (int): Link Quality Indicator of the received response.
-    """
-
-    if ack_is_disable:
-        # Cannot conclude reachability when ACK is disabled
-        return
-
-    push_APS_ACK_NACKto_plugin(self, _nwkid, Cluster, sequence, result, lqi)
-
-    if result == 0x00:
-        # Device successfully responded, remove from not reachable list
-        if _ieee in self._currently_not_reachable:
-            self._currently_not_reachable.remove(_ieee)
-
-    elif _ieee not in self._currently_not_reachable:
-        self._currently_not_reachable.append(_ieee)
-
-
-@contextlib.asynccontextmanager
-async def _limit_concurrency(self, destination, sequence):
-    """
-    Async context manager to limit concurrent requests to a specific Zigbee device.
-
-    This manager ensures that no more than a fixed number of concurrent requests
-    are issued to a single device at once, using a per-device asyncio.Semaphore.
-    If the semaphore is unavailable, the request waits up to a timeout before
-    optionally being skipped. Waiting and execution are logged for debugging.
-
-    Parameters
-    ----------
-    destination : zigpy.device.Device
-        The target Zigbee device for which concurrency is being managed.
-    sequence : int | str
-        An identifier for the request, used for logging/debugging purposes.
-
-    Yields
-    ------
-    None
-        Control is yielded once a concurrency slot is available. Code inside the
-        `async with` block executes only if the semaphore is successfully acquired.
-        If the timeout is reached before acquiring, the block is skipped and a warning
-        is logged.
-
-    Raises
-    ------
-    asyncio.TimeoutError
-        Not raised to the caller; handled internally. Timeout events are logged
-        and the block may be skipped (depending on the policy).
-
-    Notes
-    -----
-    - MAX_CONCURRENT_REQUESTS_PER_DEVICE defines the per-device concurrency limit.
-    - SEMAPHORE_TIMEOUT defines the maximum wait time to acquire a semaphore.
-    - Uses `self._concurrent_requests_semaphores_list` to track per-device semaphores.
-    - Uses `self._currently_waiting_requests_list` to track pending requests per device.
-    - Correctly handles semaphore acquisition and release to avoid race conditions.
-    - Logs the queuing, delay, and execution times for each request.
-    """
-
-    ieee = str(destination.ieee)
-    nwkid = destination.nwk.serialize()[::-1].hex()
-
-    semaphore = self._concurrent_requests_semaphores_list.setdefault(
-        ieee,
-        asyncio.Semaphore(MAX_CONCURRENT_REQUESTS_PER_DEVICE)
-    )
-
-    self._currently_waiting_requests_list.setdefault(ieee, 0)
-
-    acquired = False
-    queued = False
-    start_time = time.monotonic()
-
-    if semaphore.locked():
-        queued = True
-        self._currently_waiting_requests_list[ieee] += 1
-        self.log.logging(
-            "TransportZigpy", "Debug",
-            f"Max concurrency reached for {nwkid}, delaying request {sequence} "
-            f"({self._currently_waiting_requests_list[ieee]} enqueued)",
-            nwkid
-        )
-
-    try:
-        try:
-            await asyncio.wait_for(semaphore.acquire(), timeout=SEMAPHORE_TIMEOUT)
-            acquired = True
-        except asyncio.TimeoutError:
-            self.log.logging(
-                "TransportZigpy", "Log",
-                f"Timeout waiting for concurrency slot for {nwkid}, request {sequence} skipped",
-                nwkid
-            )
-            return  # strict mode
-
-        if queued:
-            elapsed_time = time.monotonic() - start_time
-            self.log.logging(
-                "TransportZigpy", "Debug",
-                f"Delayed request {sequence} now running after {elapsed_time:.2f}s for {nwkid}",
-                nwkid
-            )
-
-        yield
-
-    finally:
-        if acquired:
-            semaphore.release()
-
-        if queued:
-            self._currently_waiting_requests_list[ieee] = max(
-                0,
-                self._currently_waiting_requests_list[ieee] - 1
-            )
-
-def cleanup_unused_concurrency_state(self):
-    """
-    Cleans up semaphore and waiting state for inactive devices.
-
-    This method iterates through all per-device semaphores and removes entries
-    that are no longer in use. A device's concurrency state is considered unused if:
-    - No requests are currently waiting (i.e., waiting count == 0)
-    - All semaphore slots are released (i.e., the semaphore is not acquired by any task)
-
-    This helps prevent unbounded memory growth when many devices are seen temporarily.
-
-    Notes
-    -----
-    - This method should be called periodically (e.g., every hour) or during idle time.
-    - Safe to call while other coroutines are active.
-    """
-    for ieee in list(self._concurrent_requests_semaphores_list):
-        sem = self._concurrent_requests_semaphores_list[ieee]
-        waiting = self._currently_waiting_requests_list.get(ieee, 0)
-
-        # Only clean up if no one is waiting and all slots are released
-        if waiting == 0 and sem._value == MAX_CONCURRENT_REQUESTS_PER_DEVICE:
-            self.log.logging("TransportZigpy", "Debug", f"_cleanup_unused_concurrency_state {ieee} from concurrency_state", )
-            
-            del self._concurrent_requests_semaphores_list[ieee]
-            self._currently_waiting_requests_list.pop(ieee, None)
-            
-
-def specific_endpoints(self):
-    """
-    Checks if the plugin configuration enables specific endpoint handling.
-
-    Returns True if any supported plugin (Terncy, Konke, etc.) is enabled.
-
-    Returns:
-        bool: True if specific endpoints needed.
-    """
-    supported_plugins = ["Terncy", "Konke", "Wiser", "Orvibo", "Livolo", "Wiser2"]
-
-    return any(
-        plugin in self.pluginconf.pluginConf
-        and self.pluginconf.pluginConf[plugin]
-        for plugin in supported_plugins
-    )
-
-
-def _import_class(dotted_path: str):
-    """
-    Imports a class from a dotted module path string.
-
-    Example: 'Classes.ZigpyTransport.AppBellows.App_bellows'
-    → imports Classes.ZigpyTransport.AppBellows, returns App_bellows
-    """
-    module_path, class_name = dotted_path.rsplit(".", 1)
-    module = importlib.import_module(module_path)
-    return getattr(module, class_name)
